handle menu select on preview action component
